pyproject.toml
[project]
name = "portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "bandit>=1.9.4",
    "black==25.9.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.3.0",
    "dj-database-url==3.0.1",
    "django==5.1.6",
    "django-cors-headers==4.7.0",
    "django-environ==0.12.0",
    "djangorestframework==3.15.2",
    "gunicorn==23.0.0",
    "idna==3.10",
    "isort>=8.0.1",
    "mypy-extensions==1.1.0",
    "packaging==24.2",
    "pathspec==0.12.1",
    "platformdirs==4.5.0",
    "psycopg2-binary==2.9.10",
    "pytokens==0.2.0",
    "requests==2.32.3",
    "ruff>=0.15.4",
    "sqlparse==0.5.3",
    "urllib3==2.3.0",
    "whitenoise==6.9.0",
]
